backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
beautifulsoup4==4.8.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
chardet==3.0.4
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.0.8
django-cors-headers==3.13.0
djangorestframework==3.13.1
idna==2.8
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.33
pyparsing==3.0.9
pytz==2022.6
redis==4.3.5
requests==2.22.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
urllib3==1.25.11
vine==5.0.0
wcwidth==0.2.5

matplotlib~=3.6.2
numpy~=1.23.5
Pillow~=9.3.0
opencv-python~=4.6.0.66